+++ lib/main.dart
@@ -11,7 +11,7 @@ import 'package:poorlex/controller/image_picker_controller.dart';
 import 'package:poorlex/controller/layout.dart';
 import 'package:poorlex/controller/user.dart';
 import 'package:poorlex/models/login.dart';
-import 'package:poorlex/provider/battiles_provider.dart';
+import 'package:poorlex/provider/battles_provider.dart';
 import 'package:poorlex/screen/my/my_notification.dart';
 import 'package:poorlex/screen/my/my_page.dart';
 import 'package:poorlex/widget/common/webview.dart';
