+++ lib/provider/battles_provider.dart
